refactor: split xtask into modules and move task implementations out of main.rs
