+++ rs_files_over_300_loc.md
@@ -1,11 +1,11 @@
 Rust files >300 LOC (excluding target/)
-499 — xtask/src/main.rs
-471 — libs/provisioners/engine-provisioner/src/providers/llamacpp.rs
-394 — libs/provisioners/model-provisioner/src/lib.rs
-366 — bin/orchestratord/bdd/src/steps/data_plane.rs
-354 — bin/orchestratord/tests/provider_verify.rs
-326 — libs/worker-adapters/http-util/bdd/src/steps/http_util.rs
-326 — libs/catalog-core/src/lib.rs
-312 — tools/readme-index/src/render.rs
-307 — libs/worker-adapters/http-util/src/lib.rs
+[ ] 499 — xtask/src/main.rs
+[x] 471 — libs/provisioners/engine-provisioner/src/providers/llamacpp.rs
+[ ] 394 — libs/provisioners/model-provisioner/src/lib.rs
+[ ] 366 — bin/orchestratord/bdd/src/steps/data_plane.rs
+[ ] 354 — bin/orchestratord/tests/provider_verify.rs
+[ ] 326 — libs/worker-adapters/http-util/bdd/src/steps/http_util.rs
+[ ] 326 — libs/catalog-core/src/lib.rs
+[ ] 312 — tools/readme-index/src/render.rs
+[ ] 307 — libs/worker-adapters/http-util/src/lib.rs
 Summary: 9 Rust files exceed 300 lines.
